improvement: use pc for finding references of local symbols and when semanticdb is missing
