+++ presentation-compiler/test/dotty/tools/pc/utils/DefSymbolCollector.scala
@@ -7,13 +7,13 @@ import dotty.tools.dotc.ast.{Trees, tpd}
 import dotty.tools.dotc.core.Symbols.*
 import dotty.tools.dotc.interactive.InteractiveDriver
 import dotty.tools.dotc.util.SourcePosition
-import dotty.tools.pc.PcCollector
+import dotty.tools.pc.SimpleCollector
 import dotty.tools.pc.EndMarker
 
 final class DefSymbolCollector(
     driver: InteractiveDriver,
     params: VirtualFileParams
-) extends PcCollector[Option[Symbol]](driver, params):
+) extends SimpleCollector[Option[Symbol]](driver, params):
 
   def collect(parent: Option[Tree])(
       tree: Tree | EndMarker,
